+++ Source/LuaBridge/detail/ClassInfo.h
@@ -2,6 +2,7 @@
 /*
   https://github.com/vinniefalco/LuaBridge
   
+  Copyright 2020, Dmitry Tarakanov
   Copyright 2012, Vinnie Falco <vinnie.falco@gmail.com>
 
   License: The MIT License (http://www.opensource.org/licenses/mit-license.php)
@@ -29,6 +30,7 @@
 #pragma once
 
 namespace luabridge {
+namespace detail {
 
 /**
  * A unique key for a type name in a metatable.
@@ -121,49 +123,41 @@ inline const void* getParentKey ()
 #endif
 }
 
-/** Unique Lua registry keys for a class.
-
-    Each registered class inserts three keys into the registry, whose
-    values are the corresponding static, class, and const metatables. This
-    allows a quick and reliable lookup for a metatable from a template type.
+/**
+    Get the key for the static table in the Lua registry.
+    The static table holds the static data members, static properties, and
+    static member functions for a class.
 */
 template <class T>
-class ClassInfo
+void const* getStaticRegistryKey ()
 {
-public:
-  /** Get the key for the static table.
+  static char value;
+  return &value;
+}
 
-      The static table holds the static data members, static properties, and
-      static member functions for a class.
-  */
-  static void const* getStaticKey ()
-  {
-    static char value;
-    return &value;
-  }
+/** Get the key for the class table in the Lua registry.
+    The class table holds the data members, properties, and member functions
+    of a class. Read-only data and properties, and const member functions are
+    also placed here (to save a lookup in the const table).
+*/
+template<class T>
+void const* getClassRegistryKey ()
+{
+  static char value;
+  return &value;
+}
 
-  /** Get the key for the class table.
+/** Get the key for the const table in the Lua registry.
+    The const table holds read-only data members and properties, and const
+    member functions of a class.
+*/
+template<class T>
+void const* getConstRegistryKey ()
+{
+  static char value;
+  return &value;
+}
 
-      The class table holds the data members, properties, and member functions
-      of a class. Read-only data and properties, and const member functions are
-      also placed here (to save a lookup in the const table).
-  */
-  static void const* getClassKey ()
-  {
-    static char value;
-    return &value;
-  }
-
-  /** Get the key for the const table.
-
-      The const table holds read-only data members and properties, and const
-      member functions of a class.
-  */
-  static void const* getConstKey ()
-  {
-    static char value;
-    return &value;
-  }
-};
+} // namespace detail
 
 } // namespace luabridge